requirements.txt
Flask==1.1.2
numpy==1.16.6
opencv-python==4.5.1.48
Werkzeug==1.0.1
Pillow==8.1.0
scikit_learn==0.24.1
scikit-image==0.17.2
scipy==1.5.4
tensorflow-cpu==2.4.1
gunicorn